update: implement README with screenshots and tutorial structure
